requirements.txt
groq>=0.37.1          # for LLM inference
google-genai>=1.0.0       # for Gemini TTS (AI Studio API)
tenacity>=9.1.2       # for retry logic
ffmpeg-python>=0.2.0  # for video rendering
python-dotenv>=1.2.1  # for API key management
requests>=2.32.5      # for synchronous API calls (TMDB, Pexels)
openai-whisper>=20231117 # for audio transcription
pydantic>=2.0.0       # for data validation in narrative models
pysubs2>=1.7.0        # for ASS subtitle generation
Wikipedia-API>=0.6.0  # for fetching movie plots
gspread>=6.0.0        # for Google Sheets integration
google-api-python-client>=2.100.0  # for Google Drive uploads
google-auth>=2.23.0   # for service account authentication
rich>=13.0.0          # for interactive wizard UI
uvicorn>=0.44.0       # for web dashboard server
fastapi>=0.135.0      # for web dashboard API
